pyproject.toml
[project]
name = "pjgq-health-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "uvicorn>=0.38.0",
    "pyjwt>=2.8.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
